Move get-local-project-data and its test into a folder in utils/
