+++ src/components/project-overview/project-overview.test.tsx
@@ -1,9 +1,9 @@
 import ProjectOverview from '@/components/project-overview/project-overview';
+import * as getLocalProjectDataModule from '@/utils/get-local-project-data/get-local-project-data';
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
-import * as getLocalProjectDataModule from '@/utils/get-local-project-data';
 
-jest.mock('@/utils/get-local-project-data');
+jest.mock('@/utils/get-local-project-data/get-local-project-data');
 
 describe('Project Overview', () => {
   const mockProjectData = [
