+++ src/utils/get-local-project-data/get-local-project-data.tsx
